+++ packages/types/custom/core.ts
@@ -32,7 +32,7 @@ export interface PluginOptions {
   layoutTag?: string // TODO: Might it be useful to move declaration of NineLayoutTag here?
 }
 
-export type RenderType = 'iconMenu' | 'independent'
+export type RenderType = 'iconMenu' | 'independent' | 'footer'
 
 /** Possible search methods by type */
 export type SearchType = 'bkg' | 'gazetteer' | 'wfs' | 'mpapi' | string
@@ -434,6 +434,7 @@ export type ToastConfiguration = PluginOptions & ToastTypeStyles
 export interface ZoomConfiguration extends PluginOptions {
   renderType?: RenderType
   showMobile?: boolean
+  showZoomSlider?: boolean
 }
 
 export interface LanguageOption {